requirements.txt
# API Integrations
google-api-python-client==2.86.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
proxycurl==0.0.13
serpapi==0.1.2
openai==1.3.0
python-dotenv==1.0.0

# File Processing
pandas==2.2.2 # Updated from 2.0.3
openpyxl==3.1.2
python-docx==0.8.11
fpdf2==2.7.4

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.11.2
webdriver-manager==4.0.0

# API
python-multipart==0.0.6
fastapi==0.103.1
uvicorn==0.23.2
